+++ packages/core/src/utils/i18n/index.ts
@@ -17,4 +17,5 @@ export {
   type TimePickerOptionUnit
 } from '../timepicker-utils'
 export * from '../upload-labels'
-export { defineLocale, defineText } from './define-locale'
+export { defineLocale } from './define-locale'
+export { defineText } from './define-text'